+++ packages/orq-rc/src/models/operations/updateidentity.ts
@@ -167,7 +167,7 @@ export const UpdateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-12T05:19:07.172Z",
+    "2026-02-12T08:03:44.307Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {